Shifted commands ArrayList into its own class
