+++ src/com/jpii/navalbattle/NavalBattle.java
@@ -18,7 +18,7 @@
 package com.jpii.navalbattle;
 
 import com.jpii.roketgamer.RoketGamer;
-import com.jpii.navalbattle.data.Constants;
+import com.jpii.navalbattle.data.Commands;
 import com.jpii.navalbattle.data.GameState;
 import com.jpii.navalbattle.debug.CommandHandler;
 import com.jpii.navalbattle.debug.DebugWindow;
@@ -37,7 +37,7 @@ public class NavalBattle {
 		gameState = new GameState();
 		roketGamer = new RoketGamer();
 		
-		commandHandler = new CommandHandler(Constants.COMMANDS);
+		commandHandler = new CommandHandler(Commands.COMMANDS);
 		
 		debugWindow.printInfo("NavalBattle initialized.");
 		new LoginWindow();
